fix: configure cache for installation tokens
The function that gets installation tokens is cached so we don't ask for them so often.
However the cache was not being initialized.

SO the increase in requests is probably being interpreted by GitHub as rate limit violation and now we are getting 403s left and right :E

These changes configure the cache (I had to move some redis funcitons) and extends the timeout to 30minutes. THe [docs](https://docs.github.com/en/apps/creating-github-apps/authenticating-with-a-github-app/generating-an-installation-access-token-for-a-github-app) say they expire after 1h, so this should be safe to do.
