+++ shared/bots/github_apps.py
@@ -13,8 +13,8 @@ from shared.django_apps.codecov_auth.models import (
 )
 from shared.django_apps.core.models import Repository
 from shared.github import get_github_integration_token, is_installation_rate_limited
+from shared.helpers.redis import get_redis_connection
 from shared.orms.owner_helper import DjangoSQLAlchemyOwnerWrapper
-from shared.torngit.cache import get_redis_connection
 from shared.typings.oauth_token_types import Token
 from shared.typings.torngit import GithubInstallationInfo
