+++ shared/helpers/redis.py
@@ -1,0 +1,21 @@
+from redis import Redis
+
+from shared.config import get_config
+
+
+def get_redis_url() -> str:
+    url = get_config("services", "redis_url")
+    if url is not None:
+        return url
+    hostname = "redis"
+    port = 6379
+    return f"redis://{hostname}:{port}"
+
+
+def get_redis_connection() -> Redis:
+    url = get_redis_url()
+    return _get_redis_instance_from_url(url)
+
+
+def _get_redis_instance_from_url(url):
+    return Redis.from_url(url)
