+++ shared/torngit/cache/__init__.py
@@ -5,25 +5,7 @@ from typing_extensions import Literal
 
 from shared.config import get_config
 from shared.helpers.cache import OurOwnCache, RedisBackend
-
-
-def get_redis_url() -> str:
-    url = get_config("services", "redis_url")
-    if url is not None:
-        return url
-    hostname = "redis"
-    port = 6379
-    return f"redis://{hostname}:{port}"
-
-
-def get_redis_connection() -> Redis:
-    url = get_redis_url()
-    return _get_redis_instance_from_url(url)
-
-
-def _get_redis_instance_from_url(url):
-    return Redis.from_url(url)
-
+from shared.helpers.redis import get_redis_url
 
 CachedEndpoint = Union[Literal["check"], Literal["compare"], Literal["status"]]
